Remove not used nunjucks-express package
