+++ .dockerignore
@@ -15,5 +15,5 @@
 !config/**
 !webpack.config.js
 !.git
-!createVersionFile.js
+!createVersionFile.mjs
 !version
