+++ .nvmrc
@@ -1,1 +1,1 @@
-24.15.0
+24.14.1
