+++ charts/lau-frontend/Chart.yaml
@@ -3,7 +3,7 @@ appVersion: "1.0"
 description: A Helm chart for Log and Audit Frontend App
 name: lau-frontend
 home: https://github.com/hmcts/lau-frontend
-version: 0.2.16
+version: 0.2.17
 dependencies:
   - name: nodejs
     version: 3.2.1
