+++ charts/lau-frontend/values.preview.template.yaml
@@ -5,6 +5,6 @@ nodejs:
   environment:
     OAUTH_CLIENT_REDIRECT: https://${SERVICE_FQDN}/oauth2/callback
     HMCTS_ACCESS_ENABLED: true
-
+    APPLICATION_INSIGHTS_NO_STATSBEAT: true
 redis:
   enabled: true
