+++ package.json
@@ -3,12 +3,12 @@
   "version": "1.0.30",
   "private": true,
   "engines": {
-    "node": ">=24.14.0"
+    "node": ">=24.14.1"
   },
   "packageManager": "yarn@4.13.0",
   "scripts": {
     "preinstall": "./bin/preinstalls.sh || true",
-    "setup": "NODE_PATH=. node createVersionFile.js",
+    "setup": "NODE_PATH=. node createVersionFile.mjs",
     "start": "NODE_ENV=production node dist/server.js",
     "start:dev": "NODE_ENV=development nodemon --config nodemon.json",
     "start:test": "NODE_ENV=test nodemon --config nodemon.json",
@@ -65,13 +65,10 @@
     "csrf-sync": "^4.0.3",
     "email-validator": "^2.0.4",
     "express": "^5.0.0",
-    "express-nunjucks": "^3.0.0",
     "express-session": "^1.18.1",
-    "git-rev-sync": "^3.0.2",
     "govuk-frontend": "^5.0.0",
     "helmet": "^8.0.0",
     "jquery": "^4.0.0",
-    "js-yaml": "^4.1.0",
     "jwt-decode": "^4.0.0",
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
@@ -146,18 +143,6 @@
     "yarn-audit-fix": "^10.0.0"
   },
   "resolutions": {
-    "diff": "8.0.4",
-    "minimatch@^10.1.1": "10.2.4",
-    "serialize-javascript": "^7.0.3",
-    "yauzl": "^3.2.1",
-    "@react-native/codegen/glob": "^9.0.0",
-    "react-native/glob": "^9.0.0",
-    "detox/glob": "^9.0.0",
-    "rimraf@npm:2.4.5/glob": "^9.0.0",
-    "rimraf@npm:3.0.2/glob": "^9.0.0",
-    "shelljs/glob": "^9.0.0",
-    "test-exclude/glob": "^9.0.0",
-    "protobufjs": "^8.0.1",
     "uuid": "^11.1.1"
   }
 }
