--- yarn-audit-known-issues
@@ -1,1 +1,0 @@
-{"value":"glob","children":{"ID":"glob (deprecation)","Issue":"Old versions of glob are not supported, and contain widely publicized security vulnerabilities, which have been fixed in the current version. Please update. Support for old versions may be purchased (at exorbitant rates) by contacting i@izs.me","Severity":"moderate","Vulnerable Versions":"10.5.0","Tree Versions":["10.5.0"],"Dependents":["node-gyp@npm:11.1.0"]}}
